Move pool controls to PacbioPoolEdit
