+++ src/components/pacbio/PacbioPoolEdit.vue
@@ -2,7 +2,25 @@
   <b-col data-type="pool">
     <alert ref="alert" data-type="pool-create-message"></alert>
     <h3>Pooled Samples</h3>
-    <PacbioPoolLibraryList />
+    <b-row>
+      <b-col>
+        <b-form-checkbox v-model="autoTag" name="check-button" switch data-attribute="auto-tagging">
+          Autotagging
+        </b-form-checkbox>
+      </b-col>
+      <b-col>
+        <b-form-file
+          id="qcFileInput"
+          v-model="uploadedFile"
+          :state="Boolean(uploadedFile)"
+          placeholder="Choose a file or drop it here..."
+          drop-placeholder="Drop file here..."
+          accept="text/csv, .csv"
+          size="sm"
+        ></b-form-file>
+      </b-col>
+    </b-row>
+    <PacbioPoolLibraryList :auto-tag="autoTag" />
     <div class="pool-edit" data-type="pool-edit">
       <b-table-simple>
         <b-tr>
@@ -82,6 +100,8 @@
 import Alert from '@/components/Alert'
 import PacbioPoolLibraryList from '@/components/pacbio/PacbioPoolLibraryList'
 import { createNamespacedHelpers } from 'vuex'
+import { eachRecord } from '@/lib/csv/pacbio'
+
 const { mapGetters, mapActions } = createNamespacedHelpers('traction/pacbio/poolCreate')
 
 export default {
@@ -93,6 +113,8 @@ export default {
   data() {
     return {
       busy: false,
+      autoTag: false,
+      uploadedFile: null,
     }
   },
   computed: {
@@ -101,9 +123,15 @@ export default {
       return !!this.poolItem.id
     },
   },
+  watch: {
+    async uploadedFile(newFile) {
+      const csv = await newFile.text()
+      eachRecord(csv, (record) => this.updateLibraryFromCsvRecord(record))
+    },
+  },
   mounted() {},
   methods: {
-    ...mapActions(['createPool', 'updatePool']),
+    ...mapActions(['createPool', 'updatePool', 'updateLibraryFromCsvRecord']),
     create() {
       this.busy = true
       this.createPool().then(({ success, barcode, errors }) => {
